+++ database/migrations/2020_06_25_122842_create_retour_flotes_table.php
@@ -18,7 +18,7 @@ class CreateRetourFlotesTable extends Migration
             $table->integer('id_user')->unsigned()->nullable()->index();
             $table->integer('id_approvisionnement')->unsigned()->nullable()->index();
             $table->string('reference')->nullable();
-            $table->integer('montant')->unsigned()->default('0');
+            $table->double('montant')->unsigned()->default('0');
             $table->integer('reste')->unsigned()->nullable()->default('0');
             $table->string('statut')->nullable();
             $table->integer('user_destination')->unsigned()->nullable()->index();
